refactor - domain layer package 명 수정
