+++ src/main/java/gift/domain/Product.java
@@ -1,4 +1,4 @@
-package gift.entity;
+package gift.domain;
 
 public class Product {
