+++ src/main/java/gift/domain/Wish.java
@@ -1,4 +1,4 @@
-package gift.entity;
+package gift.domain;
 
 public class Wish {
     private int count;
