+++ src/main/java/gift/dto/ProductStatusPatchRequestDto.java
@@ -1,6 +1,6 @@
 package gift.dto;
 
-import gift.entity.Product;
+import gift.domain.Product;
 
 public record ProductStatusPatchRequestDto(Product.Status status) {
 }
