+++ src/main/java/gift/repository/DatabaseProductRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.entity.Product;
+import gift.domain.Product;
 import org.springframework.jdbc.core.simple.JdbcClient;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
