+++ src/main/java/gift/repository/MemberRepository.java
@@ -4,7 +4,7 @@ import org.springframework.jdbc.core.simple.JdbcClient;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import gift.entity.Member;
+import gift.domain.Member;
 import java.util.Optional;
 
 @Repository
