+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.entity.Product;
+import gift.domain.Product;
 
 import java.util.List;
 import java.util.Optional;
