+++ src/main/java/gift/service/AuthService.java
@@ -2,7 +2,7 @@ package gift.service;
 
 import gift.dto.AuthResponseDto;
 import gift.dto.MemberRequestDto;
-import gift.entity.Member;
+import gift.domain.Member;
 import gift.security.JwtTokenProvider;
 import org.springframework.stereotype.Service;
